1.8.24(II)
quick update before I call it a night 😴😴

'bvlgari.html' has been updated with the new design system. Large, bold type in front of an image will always be cool.

Next task will be to fix the spacing between the titles when I add the remainder of the body copy